.history/requirements_20250327174746.txt
torch>=2.2.0
transformer-lens @ git+https://github.com/TransformerLensOrg/TransformerLens.git@dev
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
numpy>=1.24.0
streamlit>=1.32.0
pandas>=2.2.0
requests>=2.31.0
pyyaml>=6.0.1 